+++ core/dbt/compilation.py
@@ -166,7 +166,7 @@ class Compiler:
 
     def initialize(self):
         make_directory(self.config.target_path)
-        make_directory(self.config.modules_path)
+        make_directory(self.config.packages_install_path)
 
     # creates a ModelContext which is converted to
     # a dict for jinja rendering of SQL